requirements.txt
pygithub
discord.py[voice]
dblpy
